fix: When nautilus is not installed, disable the options that requires it
Although it's forgotten to add to package dependencies. Nautilus currently is a hard-dependency of this app. If yout try to launch the app while Nautilus is not installed (spesifically nautilus-data package is not installed) it will not start and throw "Glib-GIO-ERROR **: Settings schema 'org.gnome.nautilus.icon-view' is not installed" error. Since some users may not want to use/install Nautilus i made sure that the app will launch without Nautilus installed, makes it an optional dependency and shows a warn message for the user in the UI.

Signed-off-by: Edip <edip@medip.dev>